avss: client: Improve typing of AVSSClient.reports()
Modify type hint to include the parsed report dataclasses
as a possible yield type in the generator returned by
AVSSClient.reports().

The declared type is AsyncGenerator[Report | _ParsedReport, None]
which is true, but not ideal since it implies that the generator
is heterogenous when in fact it will either yield only Report
or only _ParsedReport.