transceiver: client: Cancel connection tasks on close
The ping task and the receive tasks were not cancelled when closing the
connection by existing the TransceiverClient context manager (i.e. when
running __aexit__). This lead to the ping task trying to ping a transceiver
using a closed socket which resulted in an unhandled exception being
reported for connections that were closed expectedly and normally.

This was observable e.g. by running the transceiver upgrade CLI command.