transceiver: transport: usb: Flush old data on start, fix timeout
We now flush old data on the IN endpoint when we connect to the
device. We could alternatively reset the device, but that is more
invasive and on Linux it is even logged in dmesg.

The USB .read() function has a default timeout of 1 second, while we
want no timeout at all, so we now explicitly set it to 0 ms.