transceiver: transport: Gracefully handle missing pyusb module
Make pyusb into an optional depdendency by gracefully handling
the case wheere pyusb is missing by replacing USBTransport with
a non-functional dummy in case of a ModuleNotFoundError when
it is imported.

This lets you import and use all non-usb related features
of pyanura without requiring the pyusb dependency.

Attempting to use "usb:<SERIAL>" to create a TransceiverClient
when pyusb is missing results in a runtime error.