avss: New approach to marshalling message models
dataclasses_cbor is problematic in that customizing
the encoding/decode logic for a type relies on adding
new features to the generic conversion routine.
Increasing complexity.

This commit introduces marshal/unmarshal. Effectively
serving the same purpose but with built in support for
overriding the implemention for individual types.

The new functions have generic implementations for
dataclasses based on dataclass Fields augmented with
a cbor_key metadata attribute (just like dataclasses_cbor).

The new marshal/unmarshal functions are used in the avss client
for converting models.