marshalling: Type hint for marshal return type
Editor assistance tools like Pylance try to infer a type
for marshal but the inferred type gets weird due to
it's recursive implementation.

marshal() may return anything accepted by cbor2.dump which
supports hooks for arbitrary types so the true return type of
marshal() is Any.  That's not very helpful though so we give it
the return type `dict | list | Any` with the idea being that
we signal two things to the caller: (1) marshal() returns structured
data (dict | list) as opposed to serialized (str or bytes) (2) aside from
structured data, any other value may be returned (i.e. we let `Any`
represent all of the CBOR primitive types and any other type for
which a cbor2 hook might exist).

unmarshal is given the same typing for its input.